+ Half ton troop carrying space not working correctly

+++ megamek/src/megamek/common/verifier/TestEntity.java
@@ -777,7 +777,7 @@ public abstract class TestEntity implements TestEntityOption {
         return buff;
     }
 
-    public int getWeightCarryingSpace() {
+    public float getWeightCarryingSpace() {
         return getEntity().getTroopCarryingSpace();
     }
